requirements_v4.txt
transformers
torchaudio
scikit-learn
hdbscan
pandas
pyarrow
soundfile
tqdm
plotly
joblib
demucs
streamlit
umap-learn
pyyaml
essentia
